additional enum fix for spdlog

+++ src/device/DeviceBase.cpp
@@ -1068,7 +1068,8 @@ std::vector<StereoPair> DeviceBase::getAvailableStereoPairs() {
                         stereoPairs.push_back(dai::StereoPair{leftSocket, rightSocket, std::abs(baseline), static_cast<int>(translationVector[0]) == 0});
                     }
                 } else {
-                    pimpl->logger.debug("Skipping diagonal pair, left: {}, right: {}.", leftSocket, rightSocket);
+                    pimpl->logger.debug("Skipping diagonal pair, left: {}, right: {}.", 
+                        static_cast<std::int32_t>(leftSocket), static_cast<std::int32_t>(rightSocket));
                 }
             } catch(const std::exception&) {
                 continue;
@@ -1082,10 +1083,10 @@ std::vector<StereoPair> DeviceBase::getAvailableStereoPairs() {
     std::copy_if(
         stereoPairs.begin(), stereoPairs.end(), std::back_inserter(filteredStereoPairs), [this, connectedCameras, deviceStereoPairs](dai::StereoPair pair) {
             if(std::find(connectedCameras.begin(), connectedCameras.end(), pair.left) == connectedCameras.end()) {
-                pimpl->logger.debug("Skipping calibrated stereo pair because, camera {} was not detected.", pair.left);
+                pimpl->logger.debug("Skipping calibrated stereo pair because, camera {} was not detected.", static_cast<std::int32_t>(pair.left));
                 return false;
             } else if(std::find(connectedCameras.begin(), connectedCameras.end(), pair.right) == connectedCameras.end()) {
-                pimpl->logger.debug("Skipping calibrated stereo pair because, camera {} was not detected.", pair.right);
+                pimpl->logger.debug("Skipping calibrated stereo pair because, camera {} was not detected.", static_cast<std::int32_t>(pair.right));
                 return false;
             }
             return std::find_if(deviceStereoPairs.begin(),
